requirements.txt
### dependency
tqdm
dvc
pandas
numpy
PyYAML
tensorflow==2.5.0

### local packages -
 -e . # << uncomment this to install from setup.py
